### 2026-01-19 #9861 @crvernon
@thevolatilebit - Let me know when the following have been fixed:

> **Paper file info**:
> ```
> 🔴 Failed to discover a `Software Design` section in paper
>
> 🔴 Failed to discover a `Research Impact Statement` section in paper
>
> 🔴 Failed to discover an `AI usage disclosure` section in paper
> ```

> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ❌ MISSING DOIs
>
> - 10.1109/jproc.2002.805829 may be a valid DOI for title:  Taming heterogeneity - the Ptolemy approach
>
> ❌ INVALID DOIs
>
> - https://doi.org/10.1016/j.jmsy.2020.06.017 is INVALID because of 'https://doi.org/' prefix
> ```
