+++ paper/paper.bib
@@ -104,31 +104,6 @@
   collection = {Modelica}
 }
 
-@article{Liu2023,
-title = {Review of digital twin about concepts, technologies, and industrial applications},
-journal = {Journal of Manufacturing Systems},
-volume = {58},
-pages = {346-361},
-year = {2021},
-note = {Digital Twin towards Smart Manufacturing and Industry 4.0},
-issn = {0278-6125},
-doi = {https://doi.org/10.1016/j.jmsy.2020.06.017},
-url = {https://www.sciencedirect.com/science/article/pii/S0278612520301072},
-author = {Mengnan Liu and Shuiliang Fang and Huiyue Dong and Cunzhi Xu},
-keywords = {Digital twin, Product lifecycle, Simulation, Industrial application, Literature review},
-abstract = {Various kinds of engineering software and digitalized equipment are widely applied through the lifecycle of industrial products. As a result, massive data of different types are being produced. However, these data are hysteretic and isolated from each other, leading to low efficiency and low utilization of these valuable data. Simulation based on theoretical and static model has been a conventional and powerful tool for the verification, validation, and optimization of a system in its early planning stage, but no attention is paid to the simulation application during system run-time. With the development of new-generation information and digitalization technologies, more data can be collected, and it is time to find a way for the deep application of all these data. As a result, the concept of digital twin has aroused much concern and is developing rapidly. Dispute and discussions around concepts, paradigms, frameworks, applications, and technologies of digital twin are on the rise both in academic and industrial communities. After a complete search of several databases and careful selection according to the proposed criteria, 240 academic publications about digital twin are identified and classified. This paper conducts a comprehensive and in-depth review of these literatures to analyze digital twin from the perspective of concepts, technologies, and industrial applications. Research status, evolution of the concept, key enabling technologies of three aspects, and fifteen kinds of industrial applications in respective lifecycle phase are demonstrated in detail. Based on this, observations and future work recommendations for digital twin research are presented in the form of different lifecycle phases.}
-}
-
-@article{Grieves2014,
-  title={Digital twin: manufacturing excellence through virtual factory replication},
-  author={Grieves, Michael},
-  journal={White paper},
-  volume={1},
-  number={2014},
-  pages={1--7},
-  year={2014}
-}
-
 @inproceedings{Ledeczi2001,
   title={The generic modeling environment},
   author={Ledeczi, Akos and Maroti, Miklos and Bakay, Arpad and Karsai, Gabor and Garrett, Jason and Thomason, Charles and Nordstrom, Greg and Sprinkle, Jonathan and Volgyesi, Peter},
@@ -140,17 +115,18 @@ abstract = {Various kinds of engineering software and digitalized equipment are 
 }
 
 @article{Ptolemy,
-    author = {Johan Eker and Jorn W Janneck and Edward A. Lee
-              and Jie Liu and Xiaojun Liu and Jozsef Ludvig and
-              Sonia Sachs and Yuhong Xiong and Stephen
-              Neuendorffer},
-    title = { Taming heterogeneity - the Ptolemy approach},
-    journal = {Proceedings of the IEEE},
-    volume = {91},
-    number = {1},
-    pages = {127-144},
-    year = {2003},
-    URL = {http://chess.eecs.berkeley.edu/pubs/488.html}
+  title = {Taming heterogeneity - the Ptolemy approach},
+  volume = {91},
+  ISSN = {0018-9219},
+  url = {http://dx.doi.org/10.1109/jproc.2002.805829},
[…]
