### 2026-01-19 #9861 @crvernon
@thevolatilebit - I noticed that you section on AI usage only discusses usage as it pertains to the manuscript.  Will you also add a disclaimer for any AI usage in code development?  Thanks.
