### 2026-01-19 #9861 @thevolatilebit
@crvernon  Thank you for raising this. We had developed the framework with only inline code suggestions, which felt too limited to highlight. I have now refined the disclosure. I also consolidated three sections (Integrations, Availability and Documentation, and Application and Value Modeling Ecosystem) into a single "Research Impact Statement" with three subsections, these all relate to impact as broadly defined in the guidelines. Happy to adjust further if needed.
