### 2026-01-26 #9861 @jbytecode
@thevolatilebit  - Hi, thank you for submitting your work to JOSS. I'm the handling editor of this submission.

Do you suggest any potential reviewers for this submission? You can use the list of people link (https://reviewers.joss.theoj.org/lookup) to search people by their expertise and programming languages. Please do not use `@` character to prevent unwanted notifications when you make a list of potential reviewers if so. Thank you in advance.
