### 2026-01-31 #9861 @ajayi-joseph
@thevolatilebit Hi, I'd be happy to review this submission if you also feel its a good natch. I have years of experience in software development and have built serverless applications. My background includes production systems in healthcare and fintech.
