### 2026-01-31 #9861 @thevolatilebit
@jbytecode Thank you for taking on this submission as editor! I appreciate your contribution to the open source community.

I've identified several potential reviewers with backgrounds in Julia and agent-based/dynamical systems modeling:

- [SergeStinckwich](https://github.com/SergeStinckwich) — Serge has experience in Julia, simulation, and agent-based modeling.
- [@zhenwu0728](https://github.com/zhenwu0728) — Zhen has authored Julia packages and works in individual/agent-based modeling.
- [gchure](https://github.com/gchure) — Griffin has background in Julia, systems biology, and dynamical systems.
- [sandreza](https://github.com/sandreza) — Andre has published a Julia package and works in scientific computing and dynamical systems.

@ajayi-joseph — Thank you for volunteering! Your willingness to contribute is appreciated. That said, this submission is fairly specialized around Julia and scientific computing (dynamical systems, agent-based modeling), so I'll defer to the editor to assess fit.
