### 2026-02-01 #9861 @jbytecode
👋 @zhenwu0728, @SergeStinckwich - Would you be willing to review this submission to JOSS? You are suggested by the authors as suitable reviewers for this submission.

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:

https://joss.readthedocs.io/en/latest/review_criteria.html
