### 2026-02-08 #9861 @jbytecode
👋 @vigneshS29, @ma-sadeghi - Would you be willing to review this submission to JOSS?

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:

https://joss.readthedocs.io/en/latest/review_criteria.html
