### 2026-02-12 #9989 @editorialbot
[REVIEW]: AlgebraicAgents.jl: Hierarchical Composition of Multi-Formalism Dynamical Systems
**Submitting author:** @thevolatilebit (<a href="http://orcid.org/0009-0001-9829-6828">Jan Bí­ma</a>)
**Repository:** https://github.com/Merck/AlgebraicAgents.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.26
