### 2026-02-18 #9989 @jbytecode
@SergeStinckwich, @zhenwu0728 - Could you please start your review? Thank you in advance.
