### 2026-03-04 #9989 @zhenwu0728
I've read through the paper and tried the example of Simulating Stochastic Reaction Systems. In summary, this is a really interesting package! Here's some comments:
1. Community guidelines are missing.
2. As for the collaborative effort, all the issues and pull requests are from the authors. And the community influences on this paper is lacking in the section of software design.
3. In the research impact statement, the authors just stated that `AlgebraicAgents.jl` can be integrated with several other packages, but I don't think that's impact. The section of Applications and Value Modeling Ecosystem looks more like impact.
4. It's also necessary to state that what package can be integrated with `AlgebraicAgents.jl`. `Agents.jl`, `DifferentialEquations.jl`, and `AlgebraicDynamics.jl` are the three packages listed in the paper. Can any other package integrate with `AlgebraicAgents.jl`？It's better to state clearly the limitations.
