### 2026-03-05 b2b292e
Fix `@aagent` for world age
