+++ src/integrations/SciMLIntegration/core.jl
@@ -28,7 +28,7 @@ mutable struct DiffEqAgent <: AbstractAlgebraicAgent
     observables::Dict{Any, Int}
 
     function DiffEqAgent(name, problem::DiffEqBase.DEProblem,
-            alg = DifferentialEquations.default_algorithm(problem)[1], args...;
+            alg = DifferentialEquations.DefaultODEAlgorithm(autodiff = DifferentialEquations.AutoFiniteDiff()), args...;
             observables = Dict{Any, Int}(), kwargs...)
         problem = DifferentialEquations.remake(problem;
             p = Params(Val(DummyType), problem.p))
@@ -67,7 +67,7 @@ end
 Base.setindex!(p::Params, v, i::Int) = getfield(p, :params)[i] = v
 
 function wrap_system(name::AbstractString, problem::DiffEqBase.DEProblem, args...;
-        alg = DifferentialEquations.default_algorithm(problem)[1],
+        alg = DifferentialEquations.DefaultODEAlgorithm(autodiff = DifferentialEquations.AutoFiniteDiff()),
         kwargs...)
     DiffEqAgent(name, problem, alg, args...; kwargs...)
 end
