### 2026-03-17 37862c2
Incorporate reviewer feedback into JOSS manuscript
- Add Community Influences subsection in Software Design, situating
  composability patterns within the AlgebraicJulia ecosystem
- Add JuliaCon 2023 citation
- Extend Third-Party Integrations with explicit extensibility contract
