+++ paper/paper.bib
@@ -140,3 +140,11 @@
   publisher={ACM New York, NY},
   doi={10.1145/3617687}
 }
+
+@misc{Bima2023JuliaCon,
+  author       = {B{\'\i}ma, Jan and Ritter, Otto and Wu, Sean L.},
+  title        = {{DyVE}, a Framework for Value Dynamics},
+  year         = {2023},
+  howpublished = {Contributed talk. {JuliaCon} 2023, Massachusetts Institute of Technology, Cambridge, MA},
+  url          = {https://pretalx.com/juliacon2023/talk/DN387J/}
+}
