### 2026-03-18 #9989 @thevolatilebit
Thank you @zhenwu0728 for the review and @jbytecode for the follow-up. We have updated the paper to address the comments raised. Responses to the points follow below.

**On community guidelines:**

Thank you for raising this. We have added a `Contributing.md` file detailing how to contribute to both the framework internals and integrations with third-party packages.

**On community influences:**

We have also introduced a *Community Influences* subsection in the Software Design section, situating the package's composability patterns within the AlgebraicJulia ecosystem.

**On community adoption and research impact:**

We acknowledge that the contributor base is currently limited. We view this publication as a step toward broader visibility and adoption. The foundational ideas were presented at JuliaCon 2023 at MIT (*DyVE, a Framework for Value Dynamics*) and have since matured into a value modeling ecosystem comprising AlgebraicAgents.jl and ReactiveDynamics.jl. We have added a citation of this talk to the paper.

Regarding the Research Impact Statement: we appreciate the reviewer's observation and concur. The paper structure requires a top-level Research Impact Statement section, under which we organized three subsections — *Applications and Value Modeling Ecosystem*, *Third-Party Integrations*, and *Availability and Documentation* — each representing a distinct dimension of the framework's outreach. We agree that the first subsection most directly evidences realized impact.

**On existing integrations and compatibility with additional packages:**

The listed integrations — DifferentialEquations.jl, Agents.jl, and AlgebraicDynamics.jl — are native wrappers we developed to support our own prototyping. The framework is designed, however, to accommodate additional packages. Any third-party dynamical system framework that exposes an integrator object on a timeline with an incremental stepping interface can be wrapped as an agent by implementing two methods: `_step!` and `_projected_to`. We have extended the Integrations subsection to state this general contract.
