### 2026-03-31 #9989 @thevolatilebit
@jbytecode - Thank you for following up on this! I believe we could reach out to other potential reviewers I suggested before:

- [gchure](https://github.com/gchure) — Griffin has background in Julia, systems biology, and dynamical systems.
- [sandreza](https://github.com/sandreza) — Andre has published a Julia package and works in scientific computing and dynamical systems.
