### 2026-03-31 #9989 @jbytecode
@gchure - One of our reviewers has completed their review. We failed to get a response from the other one. I need a reviewer to check out this submission. Would you be willing to review this submission for JOSS?
