### 2026-04-02 #9989 @jbytecode
@mkitti - One of our reviewers has completed their review. We failed to get a response from the other one. I need a reviewer to check out this submission. Would you be willing to review this submission for JOSS?
