### 2026-05-05 #9989 @mkitti
> **License info**:
>
> 🟡 License found: `Other` ([Check here](https://opensource.org/licenses) for OSI approval)

The license is a MIT License:
https://github.com/Merck/AlgebraicAgents.jl/blob/main/LICENSE.txt

Other licenses are listed for dependencies, which are also mostly MIT License as well:
https://github.com/Merck/AlgebraicAgents.jl/blob/main/LICENSES_THIRD_PARTY.txt
