### 2026-05-09 d2d0d8a
Address JOSS reviewer feedback on docs, README, paper, and license
- Relicense with canonical MIT template so GitHub/JOSS licensee
  auto-detects MIT (curly quotes, extra attribution line, and "All
  rights reserved" in the previous text broke the text match)
- Add Installation section with Pkg.jl pointer to README, docs, and
  paper; note package is registered in General registry
- State intended audience in paper, README, and docs (users coupling
  SciML/Agents.jl/AlgebraicDynamics.jl models and practitioners
  wrapping models outside these frameworks; applied-modeling focus)
- Make README "
[…]
