+++ docs/src/index.md
@@ -1,238 +1,89 @@
-# API Documentation
+# AlgebraicAgents.jl
 
-```@meta
-CurrentModule = AlgebraicAgents
+*A Julia framework for the hierarchical, heterogeneous co-integration of
+dynamical systems.*
+
+AlgebraicAgents.jl lets you compose differential equations, discrete-event systems, and agent-based models within a single rooted hierarchy. Subsystems evolve at their own pace — the framework keeps them in sync through a minimal stepping interface — and may freely observe or interact with one another.
+
+- **Multi-formalism.** Couple ODEs/SDEs, discrete-event systems, and
+  agent-based models in a single simulation.
+- **Hierarchical.** Organize subsystems into a tree, reference them by
+  filesystem-like paths, and compose them with a structural sum `⊕`.
+- **Extensible.** Wrap a custom solver or a third-party integrator as an agent
+  by implementing two interface methods: [`AlgebraicAgents._step!`](@ref) and
+  [`AlgebraicAgents._projected_to`](@ref).
+- **Introspective.** Declare information flows between agents as *wires*,
+  associate agents with atemporal *concepts* and typed *relations*, and
+  visualize the resulting architecture.
+
+## Installation
+
+AlgebraicAgents.jl is registered in the Julia General registry, so it can be installed with Julia's built-in package manager [Pkg.jl](https://pkgdocs.julialang.org/v1/managing-packages/#Adding-packages):
+
+```julia
+using Pkg
+Pkg.add("AlgebraicAgents")
 ```
 
-## Agent types
+Third-party integrations (DifferentialEquations.jl, Agents.jl,
+AlgebraicDynamics.jl) load automatically once the corresponding package is
+available in your environment — see [Integrations](integrations.md).
 
-```@docs
-AbstractAlgebraicAgent
-FreeAgent
-FreeAgent(::AbstractString, ::Vector{<:AbstractAlgebraicAgent})
-```
+## Intended Audience
 
-## Implementing custom types
+AlgebraicAgents.jl is intended for Julia modelers composing heterogeneous
+dynamical systems across modeling formalisms — users of ecosystems and
+frameworks such as SciML, Agents.jl, and AlgebraicDynamics.jl coupling
+existing models, and practitioners wrapping models built outside these
+frameworks. Primary applications lie in pharmaceutical value-chain modeling,
+systems biology, and multi-physics engineering.
 
-To implement a custom agent type, you may want to use the convenience macro [`@aagent`](@ref) which supplies type fields expected (not required, though) by the interface.
+## A Minimal Example
 
-### Required methods
+A custom agent type is declared with the [`@aagent`](@ref) macro and given an
+evolution rule by extending [`AlgebraicAgents._step!`](@ref):
 
-```@docs
-AlgebraicAgents._step!(::AbstractAlgebraicAgent)
-AlgebraicAgents._projected_to(::AbstractAlgebraicAgent)
-```
-
-### Optional methods
-
-```@docs
-AlgebraicAgents._getparameters(::AbstractAlgebraicAgent)
-AlgebraicAgents._setparameters!(::AbstractAlgebraicAgent, ::Any)
-AlgebraicAgents._draw(::AbstractAlgebraicAgent)
-AlgebraicAgents._reinit!(::AbstractAlgebraicAgent)
-AlgebraicAgents._interact!(::AbstractAlgebraicAgent)
-AlgebraicAgents._prestep!(::AbstractAlgebraicAgent, ::Float64)
-```
-
-Other optional methods include
[…]
