+++ CONTRIBUTING.md
@@ -2,9 +2,9 @@ Thanks for taking the time to contribute — we appreciate it very much!
 
 We welcome contributions to both the core interface of AlgebraicAgents.jl, refining and extending it for improved user experience and modeling expressivity, and integrations with third-party Julia packages, extending the set of formalisms that can be accommodated within the framework and composed with each other.
 
-Please see the [online documentation](link) for core design principles and internal details. For integrating third-party packages, consult the documentation on [required and optional interface methods for custom agent types](https://github.com/openjournals/joss-reviews/issues/9989#issuecomment-3996072845), and review the [existing integration source code](https://github.com/Merck/AlgebraicAgents.jl/tree/main/src/integrations) for examples of how to write wrappers for third-party simulation objects, including model stepping, state accessors, or plotting functions.
+Please see the [online documentation](https://merck.github.io/AlgebraicAgents.jl/stable/) for core design principles and internal details. For integrating third-party packages, consult the documentation on [required and optional interface methods for custom agent types](https://merck.github.io/AlgebraicAgents.jl/stable/api.html#Implementing-custom-types), and review the [existing integration source code](https://github.com/Merck/AlgebraicAgents.jl/tree/main/src/integrations) for examples of how to write wrappers for third-party simulation objects, including model stepping, state accessors, or plotting functions.
 
 Contributions to AlgebraicAgents.jl are welcome in the following forms:
 
 - Modifying the code or documentation via a pull request.
-- Reporting bugs or suggesting enhancements in the project's [GitHub Issues](link). If you propose a feature for future development, we are happy to discuss and take on the implementation.
+- Reporting bugs or suggesting enhancements in the project's [GitHub Issues](https://github.com/Merck/AlgebraicAgents.jl/issues). If you propose a feature for future development, we are happy to discuss and take on the implementation.

[…]
