### 2026-05-09 de45652
Switch code formatter from JuliaFormatter (SciML style) to Runic
- Replace .github/workflows/Formatter.yml with Format.yml mirroring
  CEEDesigns.jl: runs fredrikekre/runic-action@v1 on push to main and
  on pull requests
- Remove .JuliaFormatter.toml; Runic is configuration-free
- Apply Runic formatting across src, test, tutorials, and docs

Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
