+++ tutorials/sciml/sciml.jl
@@ -8,10 +8,10 @@ using AlgebraicAgents
 using DifferentialEquations
 
 ## vanilla function
-f(u,p,t) = 1.01*u
-u0 = 1/2
-tspan = (0.0,10.0)
-prob = ODEProblem(f,u0,tspan)
+f(u, p, t) = 1.01 * u
+u0 = 1 / 2
+tspan = (0.0, 10.0)
+prob = ODEProblem(f, u0, tspan)
 
 # ## Atomic Models
 
@@ -26,16 +26,16 @@ push!(observables(m3), "o1" => 1, "o2" => 1)
 ## simple function, calls to which will be scheduled during the model integration
 custom_function(agent, t) = println("inside $agent at time $t")
 
-## a bit more intricate logic - 
-function f_(u,p,t)
+## a bit more intricate logic -
+function f_(u, p, t)
     ## access the wrapping agent (hierarchy bond)
     agent = extract_agent(p)
-    
-    ## access observables 
+
+    ## access observables
     o1 = getobservable(getagent(agent, "../model3"), "o1")
     o2 = getobservable(getagent(agent, "../model3"), "o2")
     ## fetch observable's value at **a given time point in the past**
-    o3 = gettimeobservable(getagent(agent, "../model3"), t/2, 1)
+    o3 = gettimeobservable(getagent(agent, "../model3"), t / 2, 1)
 
     ## schedule interaction
     ## first, schedule a call to `_interact!(agent)` with priority 0
@@ -45,16 +45,16 @@ function f_(u,p,t)
     ## this will be expanded to a call f(agent, args...)
     @call agent custom_function(agent, t)
 
-    min(2., 1.01*u + o1 + o2 + o3)
+    return min(2.0, 1.01 * u + o1 + o2 + o3)
 end
 
 # ## Another Atomic Model
 
-m4 = DiffEqAgent("model4", ODEProblem(f_,u0,tspan))
+m4 = DiffEqAgent("model4", ODEProblem(f_, u0, tspan))
 
 # ## Hierarchical Sum of Atomic Models
 
-m = ⊕(m1, m2; name="diagram1") ⊕ ⊕(m3, m4; name="diagram2")
+m = ⊕(m1, m2; name = "diagram1") ⊕ ⊕(m3, m4; name = "diagram2")
 
 #
 

[…]
