### 2026-05-10 #9989 @mkitti
I may still have a few more thoughts.  I will write them today if they come to mind.
