### 2026-05-11 1c2e382
Fix SciML integration for DifferentialEquations.jl reshuffle
`DifferentialEquations` no longer re-exports `DiffEqBase` or `AutoFiniteDiff`,
which broke Requires-time loading of `DiffEqAgent` and failed the SciML test
set on the entire CI matrix. Switch to `SciMLBase` (which exposes
`DEIntegrator`, `DEProblem`, `init`, `step!`) and reach `AutoFiniteDiff`
through `SciMLBase.ADTypes`.

Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
