### 2026-05-12 #9989 @thevolatilebit
@jbytecode — we've addressed all the points raised, and the changes are consolidated in [this commit](https://github.com/Merck/AlgebraicAgents.jl/commit/d2d0d8ad9106a6deb1b6b3fee5bdfa1be59c73a4). Summary:

- **Intended audience**: Stated explicitly in the paper, README, and docs — practitioners coupling SciML, Agents.jl, and AlgebraicDynamics.jl models, plus those wrapping models from outside these frameworks; applied-modeling focus.
- **Installation**: Added an Installation section pointing to Pkg.jl in the README, docs, and paper, and noted that the package is registered in the General registry.
- **Documentation**: New landing page at `index.md`, API reference moved to `api.md`, and `CONTRIBUTING.md` staged as the Contributing page at build time.
- **README example**: Made "A Single Sketch" self-contained with the necessary `using` statements, abstract type declarations, and `const` preamble.
- **License**: Replaced `LICENSE` with the canonical MIT template so GitHub's `licensee` (and JOSS) now auto-detect MIT. The previous text failed the match due to curly quotes, an extra attribution line, and an "All rights reserved" clause.
