### 2026-05-14 95949d2
Fix tutorials and add Manifest.toml for reproducibility
- Fix SciMLIntegration: SciMLBase.DEProblem renamed to AbstractDEProblem in SciMLBase v3
- Fix molecules tutorial: EM() no longer re-exported by DifferentialEquations v8; add explicit StochasticDiffEq dep
- Add relations/Project.toml (was missing entirely)
- Add Manifest.toml for sciml, molecules, traces, stochastic_simulation, algebraicdynamics, relations tutorials
- Update .gitignore: change Manifest.toml to /Manifest.toml so tutorial Manifests can be tracked

All Manifest.toml files pin AlgebraicAgents to the local dev path (path = "..
[…]
