### 2026-05-15 b2cbc3b
chore: incorporate tutorial Manifests + fixes from #45
Apply mkitti's #45 changes that complement #44 and 1c2e382:
- Track Manifest.toml for sciml, molecules, traces, stochastic_simulation,
  algebraicdynamics, relations tutorials so reviewers can reproduce builds
- Restrict the Manifest.toml gitignore to the repo root and ignore the
  graph5.svg artifact emitted by the wires tutorial
- Add the missing tutorials/relations/Project.toml
- Make the molecules tutorial work after the DifferentialEquations v8
  reshuffle by importing EM from StochasticDiffEq directly
- Sharpen the SciML wrap to SciM
[…]
