+++ tutorials/relations/Project.toml
@@ -1,0 +1,2 @@
+[deps]
+AlgebraicAgents = "f6eb0ae3-10fa-40e6-88dd-9006ba45093a"
