### 2026-05-15 9f12cdf
docs(paper): address mkitti's minor points on the JOSS manuscript
- State of the Field: cite Catalyst.jl and explicitly contrast its
  symbolic-subsystem composition with the black-box, non-symbolic
  composition AlgebraicAgents.jl provides; distinguish from Agents.jl
  by name vs. role.
- Synchronized Evolution: add a caption to the frontier-stepping table
  (clarifies that table cells are projected times after the step) and
  rework the step! sketch so it no longer relies on minimum(...; init=nothing),
  with a note that production code uses @ret.
- Figure 1: rewrite caption to match the ren
[…]
