+++ paper/paper.md
@@ -43,9 +43,9 @@ AlgebraicAgents.jl is intended for Julia modelers composing heterogeneous dynami
 
 Prior work addresses aspects of this challenge. The Functional Mock-up Interface [@Blochwitz2011] standardizes co-simulation of black-box models but imposes protocol overhead suited to industrial interoperability rather than rapid prototyping. Meta-modeling frameworks like the Generic Modeling Environment [@Ledeczi2001] operate at a higher abstraction, enabling construction of domain-specific formalisms. The Ptolemy project [@Ptolemy] and Lingua Franca [@Menard2023] provide principled foundations for heterogeneous component interaction across concurrent, real-time, and distributed settings.
 
-Within the Julia ecosystem [@Julia2017], ModelingToolkit.jl [@Ma2021; @DifferentialEquations2017], and its commercial extension, JuliaHub's Dyad, excel at symbolic-numeric modeling and equation-based composition but do not naturally accommodate discrete or agent-based dynamics. AlgebraicDynamics.jl [@Brown2022; @Baez2023] brings categorical semantics to dynamical systems yet enforces strict interface typing that constrains exploratory work.
+Within the Julia ecosystem [@Julia2017], ModelingToolkit.jl [@Ma2021; @DifferentialEquations2017] and its commercial extension, JuliaHub's Dyad, target symbolic-numeric modeling and equation-based composition. Catalyst.jl [@Catalyst2023] builds on ModelingToolkit.jl with hierarchical subsystem composition for chemical reaction networks. All require components expressible as equations or as a symbolic intermediate representation, which limits their ability to accommodate discrete or agent-based dynamics. AlgebraicDynamics.jl [@Brown2022; @Baez2023] supplies categorical semantics for dynamical systems but imposes strict interface typing that constrains exploratory work. Agents.jl [@Agents2022], despite the similar name, solves a complementary problem: it provides a performant runtime for individual agent-based models, whereas AlgebraicAgents.jl coordinates such models alongside other formalisms within a hierarchy.
 
-AlgebraicAgents.jl relaxes formalism and shifts focus towards compositional flexibility. Any agent can access any other agent's state, and synchronization is temporal rather than type-enforced. This design prioritizes iteration speed and introspection over type-enforced interface contracts, a trade-off suited to exploratory modeling, where specifications evolve alongside understanding.
+AlgebraicAgents.jl relaxes these requirements in favor of compositional flexibility. Components are black boxes: no equation, symbolic representation, or typed interface contract is required; an agent need only expose an internal clock and an incremental stepping rule. Any agent can access any other agent's state, and synchronization is temporal rather than type-enforced. The design prioritizes iteration speed and introspection over interface contracts, a trade-off suited to exploratory modeling, where specifications evolve alongside understanding.
 
 # Software Design
 
@@ -75,17 +75,22 @@ This mechanism is exemplified below for the case of three agents A, B, and C.
 | 4           | **3**          | 3                | 3              | 3              | A       |
 | 5           | **4**          | **4.5**          | **6**          | 4              | A, B, C |
 
-Formally, the single step of the simulator is defined as follows.
+Table: Stepping mechanism for three agents with step sizes Δt = 1, 1.5, 3. Columns 2--4 show each agent's projected time after each simulator step. Bold entries mark the agents that advanced in that step. Only agents at the minimum projected time—the frontier—step, and the rest remain projected ahead and are read when queried.
+
+Pseudocode for a single simulator step is sketched below.
 
 ```julia
 function step!(a::Agent, t=projected_to(a))
     # Recurse depth-first.
-    t_min = minimum(step!(c, t) for c in children(a); init=nothing)
-    
+    t_min = minimum(
[…]
