+++ paper/paper.bib
@@ -54,15 +54,15 @@
 @article{Agents2022,
   doi = {10.1177/00375497211068820},
   url = {https://doi.org/10.1177/00375497211068820},
-  year = {2022},
-  month = jan,
+  year = {2024},
+  month = oct,
   publisher = {{SAGE} Publications},
-  pages = {003754972110688},
+  pages = {1019--1031},
   author = {George Datseris and Ali R. Vahdati and Timothy C. DuBois},
   title = {Agents.jl: a performant and feature-full agent-based modeling software of minimal code complexity},
   journal = {{SIMULATION}},
-  volume = {0},
-  number = {0},
+  volume = {100},
+  number = {10},
 }
 
 @misc{Brown2022,
