### 2026-05-18 ee35767
refactor: migrate from Requires.jl to package extensions
Optional integrations (DifferentialEquations, Agents, AlgebraicDynamics,
DataFrames, Plots, Graphviz_jll) now load via Julia 1.9+ package
extensions in `ext/` instead of `Requires.@require` in `src/requires.jl`.

Type declarations live in the main module (`src/integrations/{sciml,
agents,algebraicdynamics,draw_df}.jl`) so users can reference
`DiffEqAgent`, `ABMAgent`, `AAgent`, `GraphicalAgent` without first
loading the corresponding optional dependency. Constructors, stepping,
plotting, and pretty-printing are provided by extensions tha
[…]
