+++ src/integrations/SciMLIntegration/core.jl
@@ -1,160 +1,0 @@
-import .DifferentialEquations
-import .DifferentialEquations: SciMLBase
-
-# wrap problem, integrator, solution type; DiffEq agents supertype
-export DiffEqAgent
-
-# define DE algebraic wrap
-"""
-    DiffEqAgent(name, problem[, alg]; observables=nothing, kwargs...)
-Initialize DE problem algebraic wrap. 
-
-# Keywords
-- `observables`: either `nothing` or a dictionary which maps keys to observable's positional index in `u`,
-- other kwargs will be passed to the integrator during initialization step.
-"""
-mutable struct DiffEqAgent <: AbstractAlgebraicAgent
-    uuid::UUID
-    name::AbstractString
-
-    parent::Union{AbstractAlgebraicAgent, Nothing}
-    inners::Dict{String, AbstractAlgebraicAgent}
-
-    relpathrefs::Dict{AbstractString, UUID}
-    opera::Opera
-
-    integrator::SciMLBase.DEIntegrator
-
-    observables::Dict{Any, Int}
-
-    function DiffEqAgent(
-            name, problem::SciMLBase.AbstractDEProblem,
-            alg = DifferentialEquations.DefaultODEAlgorithm(autodiff = SciMLBase.ADTypes.AutoFiniteDiff()), args...;
-            observables = Dict{Any, Int}(), kwargs...
-        )
-        problem = DifferentialEquations.remake(
-            problem;
-            p = Params(Val(DummyType), problem.p)
-        )
-
-        # initialize wrap
-        i = new()
-        setup_agent!(i, name)
-
-        i.integrator = SciMLBase.init(problem, alg, args...; kwargs...)
-        i.observables = observables
-        i.integrator.p.agent = i
-
-        return i
-    end
-end
-
-## params wrap
-mutable struct Params
-    agent::Any
-    params::Any
-end
-
-# property indexing
-function Base.getproperty(p::Params, k::Symbol)
-    return k == :agent ? getfield(p, :agent) : getproperty(getfield(p, :params), k)
-end
-
-Base.propertynames(p::Params) = propertynames(getfield(p, :params))
-
-# vector interface
-Base.length(p::Params) = length(getfield(p, :params))
-Base.getindex(p::Params, i::Int) = getfield(p, :params)[i]
-function Base.getindex(::Params, ::Any)
-    return @error "please pass, and index, params as a vector, see https://github.com/SciML/SciMLBase.jl/pull/262"
-end
-Base.setindex!(p::Params, v, i::Int) = getfield(p, :params)[i] = v
-
-function wrap_system(
-        name::AbstractString, problem::SciMLBase.AbstractDEProblem, args...;
-        alg = DifferentialEquations.DefaultODEAlgorithm(autodiff = SciMLBase.ADTypes.AutoFiniteDiff()),
-        kwargs...
-    )
-    return DiffEqAgent(name, problem, alg, args...; kwargs...)
-end
-
-# implement common interface
-function getobservable_index(a::DiffEqAgent, obs)
-    return get(observables(a), obs, obs)
-end
-
-function getobservable(a::DiffEqAgent, obs)
-    return a.integrator.u[getobservable_index(a, obs)]
-end
-
-function gettimeobservable(a::DiffEqAgent, t::Float64, obs)
-    return a.integrator(t)[getobservable_index(a, obs)]
-end
-
-# implement internal step function
-function _step!(a::DiffEqAgent)
-    ret = SciMLBase.step!(a.integrator)
-    return ret == true && return true
-end
-
-function _getparameters(a::DiffEqAgent)
-    p = getfield(a.integrator.p, :params)
-
[…]
