### 2026-05-19 #9989 @thevolatilebit
@mkitti Thanks again for the review and all the suggestions.

@jbytecode Thanks for following up. I am also doing a final read now. Also, how would you recommend we set the article date — is it populated automatically at publication?
