### 2026-05-19 c1fe6d2
docs(paper): final pre-submission polish
- Add ORCIDs for Bíma and Wu in the author block.
- Tighten Summary parallelism (annotating / visualizing / querying).
- Minor grammar: comma after AlgebraicDynamics.jl, `its` similar name,
  rephrase the stepping illustration and table caption, replace
  "evolutionary step" with "stepping update".
- Correct the SciML wrapping description: DiffEqAgent wraps
  DifferentialEquations.jl problems (`SciMLBase.AbstractDEProblem`);
  `DEProblem` was removed in SciMLBase v3.
- Point the Contributing link to CONTRIBUTING.md on `main` instead of
  the README's co
[…]
