### 2026-05-19 1a682d3
fix(docs): unblock Documenter ExpandTemplates from agents tutorial readline
Two issues were combining to time out the Documentation workflow at the
6h job limit:

* `tutorials/agents/agents.jl` ended with `if !isinteractive() readline()
  end`. Documenter runs the Literate-generated `@example` blocks with
  `isinteractive() == false`, so the readline blocked on stdin forever
  during ExpandTemplates and the runner sat idle until cancelled.
* The `wrap_system` `@example` in `docs/src/api.md` failed because the
  SciML extension reached for `SciMLBase.ADTypes`, which no longer
  exists; ADTypes 
[…]
