### 2026-05-19 01702d4
docs(paper): drop hardcoded submission date from front matter
JOSS sets the accepted date itself; carrying a stale `date:` here just
risks publishing the wrong one.

Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
