### 2026-05-20 aa20f8e
docs: explain package-extension loading mechanism
Add a "How integrations are loaded" section to integrations.md describing
the [weakdeps]/[extensions] mechanism, the types-in-main / methods-in-ext
split, and multi-trigger combinations. Collapse the duplicate explanation
in api.md into a pointer at the new section, and expand the README's
installation note to mention extensions and link to the docs.

Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
