### 2026-05-20 50ddf31
docs: refine wording in integrations loading section
Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
