+++ versions.js
@@ -1,7 +1,8 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.4",
   "v0.3",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.3.27";
+var DOCUMENTER_NEWEST = "v0.4.0";
 var DOCUMENTER_STABLE = "stable";
