+++ v0.4.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"<script type=\"module\">\n  import mermaid from 'https://cdn.jsdelivr.net/npm/mermaid@9/dist/mermaid.esm.min.mjs';\n  mermaid.initialize({ startOnLoad: true });\n</script>","category":"page"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"<script type=\"module\">\n  import mermaid from 'https://cdn.jsdelivr.net/npm/mermaid@9/dist/mermaid.esm.min.mjs';\n  mermaid.initialize({ startOnLoad: true });\n</script>","category":"page"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"<script type=\"module\">\n  import mermaid from 'https://cdn.jsdelivr.net/npm/mermaid@9/dist/mermaid.esm.min.mjs';\n  mermaid.initialize({ startOnLoad: true });\n</script>","category":"page"},{"location":"design_mmd.html#Framework-design","page":"Framework design","title":"Framework design","text":"","category":"section"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"Here we describe the design principles of the AlgebraicAgents. It should be of most use to advanced users and persons interested in contributing to the software. New users are encouraged to start by reading one of the tutorials (\"sketches\").","category":"page"},{"location":"design_mmd.html#Simulation-loop","page":"Framework design","title":"Simulation loop","text":"","category":"section"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"We describe here the main simulation loop which steps models built in AlgebraicAgents forward in time.","category":"page"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"AlgebraicAgents keeps agents synchronized by ensuring that the model will only simulate the agent(s) whose projected time (e.g. the maximum time for which that agent's trajectory has been solved) is the minimum of all agents. For example, if there are 3 agents, whose internal step sizes are of 1, 1.5, and 3 time units, respectively, then at the end of the first step, their projected time will be 1, 1.5, and 3 (assuming they all start at time 0). The simulation will find the minimum of those times and only simulate the agent(s) whose projected time(s) are equal to the minimum. In this case, it is the first agent, who is now projected to time 2 on the second step (other agents are not simulated). Now, the new minimum time is 1.5, and the second agent is projected to time 3 on the third step, etc. The simulation continues until all agents have been projected to their final time point, or the minimum of projected times reaches a maximum time horizon. If all agents take time steps of the same size, then they will all be updated each global step.","category":"page"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"There are several functions in the interface for an AbstractAlgebraicAgent which implement these dynamics. When defining new agent types, one should implement the AlgebraicAgents._step! method, which will step that agent forward if its projected time is equal to the least projected time, among all agents in the hierarchy. Agent types also need to implement AlgebraicAgents._projected_to, which is crucial to keeping the simulation synchronized. It will return:","category":"page"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"nothing if the agent does not implement its own _step! rule (e.g. FreeAgent which is a container of agents)\ntrue if the agent has been projected to the final time point (_step! will not be called again)\na value of Number, giving the time point to which the agent has been projected","category":"page"},{"location":"design_mmd.html","page":"Framework design","title":"Framework design","text":"These are collected into ret, which is an object that will be true if and only 
[…]
