### 2026-05-24 58f94a6
Minor fix on paper and bibliography
