+++ paper/paper.bib
@@ -30,7 +30,7 @@
  note = {Exported from https://app.dimensions.ai on 2019/05/05},
  number = {1},
  pages = {},
- title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+ title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in {J}ulia},
  url = {https://app.dimensions.ai/details/publication/pub.1085583166 and http://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/galley/245/download/},
  volume = {5},
  year = {2017}
@@ -130,7 +130,7 @@
 }
 
 @article{Menard2023,
-  title={High-performance deterministic concurrency using Lingua Franca},
+  title={High-performance deterministic concurrency using {L}ingua {F}ranca},
   author={Menard, Christian and Lohstroh, Marten and Bateni, Soroush and Chorlian, Matthew and Deng, Arthur and Donovan, Peter and Fournier, Cl{\'e}ment and Lin, Shaokai and Suchert, Felix and Tanneberger, Tassilo and others},
   journal={ACM Transactions on Architecture and Code Optimization},
   volume={20},
