### 2026-05-24 36a5ff4
Minor fix on paper and bibliography (#46)
