### 2026-05-25 a36a1b4
paper: sharpen wires paragraph in Software Design
- Replace "execution hierarchy" with "containment hierarchy"
- Reframe wires as "first-class annotations" rather than
  "documentation"
