### 2026-06-25 72ef864
docs(paper): address JOSS review paper-level points (#9989)
- Use "value-chain" hyphenation consistently
- Cite companion packages ReactiveDynamics.jl and CEEDesigns.jl
- Expand ODE/SDE/DDE/DAE acronyms on first use
- Protect "Ptolemy" capitalization in bib title
- Fix Catalyst.jl author field (uppercase AND broke BibTeX parsing)

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
