+++ paper/paper.bib
@@ -38,7 +38,7 @@
 
 @article{Catalyst2023,
     doi = {10.1371/journal.pcbi.1011530},
-    author = {Loman, Torkel E. AND Ma, Yingbo AND Ilin, Vasily AND Gowda, Shashi AND Korsbo, Niklas AND Yewale, Nikhil AND Rackauckas, Chris AND Isaacson, Samuel A.},
+    author = {Loman, Torkel E. and Ma, Yingbo and Ilin, Vasily and Gowda, Shashi and Korsbo, Niklas and Yewale, Nikhil and Rackauckas, Chris and Isaacson, Samuel A.},
     journal = {PLOS Computational Biology},
     publisher = {Public Library of Science},
     title = {Catalyst: Fast and flexible modeling of reaction networks},
@@ -115,7 +115,7 @@
 }
 
 @article{Ptolemy,
-  title = {Taming heterogeneity - the Ptolemy approach},
+  title = {Taming heterogeneity - the {P}tolemy approach},
   volume = {91},
   ISSN = {0018-9219},
   url = {http://dx.doi.org/10.1109/jproc.2002.805829},
@@ -147,4 +147,18 @@
   year         = {2023},
   howpublished = {Contributed talk. {JuliaCon} 2023, Massachusetts Institute of Technology, Cambridge, MA},
   url          = {https://pretalx.com/juliacon2023/talk/DN387J/}
+}
+
+@misc{ReactiveDynamics,
+  author       = {B{\'\i}ma, Jan and Ritter, Otto and Wu, Sean L.},
+  title        = {{ReactiveDynamics.jl}: A modeling package for reaction network-type dynamical systems},
+  year         = {2022},
+  howpublished = {\url{https://github.com/Merck/ReactiveDynamics.jl}}
+}
+
+@misc{CEEDesigns,
+  author       = {B{\'\i}ma, Jan and Chen, Tianchi and Dadashova, Kamala and Jones, Gwenyth and Ritter, Otto and Wu, Sean L.},
+  title        = {{CEEDesigns.jl}: Cost-efficient experimental design in {J}ulia},
+  year         = {2023},
+  howpublished = {\url{https://github.com/Merck/CEEDesigns.jl}}
 }
