### 2026-06-25 e837b8f
fix: bump Agents.jl compat to v7 and fix macro hygiene regression
Agents.jl v6 no longer precompiles against current Distributions.jl (a
`@check_args` macro incompatibility), breaking CI. Bump the compat bound
to `Agents = "7"`. None of v7's breaking changes (plotting rewrite, OSM
spaces moved to an extension, route functions namespaced, `ByKind`
scheduler removed) touch the API the integration uses.

While verifying against the Agents tutorial, surfaced a latent macro
hygiene bug from the Requires->extensions migration (ee35767): `@get_model`
and `@a` expanded to a bare `Agents.abmproperties(
[…]
