+++ test/integrations/agents_test.jl
@@ -61,3 +61,35 @@ end
 
     @test draw(abm_algebraic_wrap) isa Plots.Plot
 end
+
+# Exercise the `@get_model` and `@a` macros from within a *stepped* model. These
+# macros expand to references that must resolve against the loaded Agents
+# extension rather than the `Agents` weakdep symbol inside `AlgebraicAgents`;
+# wiring them into `agent_step!` guards against that regression.
+function macro_agent_step!(agent, model)
+    @get_model model
+    extract_agent(model, agent)
+    # remove every infected agent via the algebraic `@a` wrapper
+    return if agent.status == :I
+        @a remove_agent!(agent, model)
+    end
+end
+
+@testset "`@get_model` and `@a` macros within stepping" begin
+    Random.seed!(1)
+    space = GraphSpace(Agents.Graphs.complete_graph(4))
+    abm_macro = StandardABM(
+        PoorSoul, space; agent_step! = macro_agent_step!, model_step! = identity,
+        properties = Dict{Symbol, Any}()
+    )
+    for _ in 1:20
+        add_agent!(rand(1:4), abm_macro, 0, :I)
+    end
+
+    abm_macro_wrap = ABMAgent("sir_macro_model", abm_macro; tspan = (0.0, 5.0))
+    # both macros are evaluated during stepping; this should not error
+    @test simulate(abm_macro_wrap) isa AlgebraicAgents.ABMAgent
+    # `@a remove_agent!` keeps the algebraic wrap's inners in sync with the ABM and
+    # removes the infected agents
+    @test length(inners(abm_macro_wrap)) == Agents.nagents(abm_macro_wrap.abm) == 0
+end
