### 2026-06-26 728b60b
fix: repair `@a add_agent!` branch referencing undefined `ABAModel`
The `@a add_agent!` branch expanded to `entangle!(omodel, ABAModel(string(agent.id), a))`,
but `ABAModel` was never defined in any commit and `a` was an undefined free
variable — the branch errored on every use. It was never caught because no test
or tutorial exercised `@a add_agent!` (only `@a remove_agent!`).

The intended wrap is `AAgent`, the algebraic wrap for an `Agents.AbstractAgent`,
matching what the `ABMAgent` constructor entangles for each existing agent
(`AAgent(string(id))`). Wrap the newly-added agent the same wa
[…]
