### 2026-06-26 72432f0
Add DOI badge to README (#47)
